+++ spec/query_methods/inet_query_spec.rb
@@ -3,7 +3,7 @@
 require "spec_helper"
 
 RSpec.describe "Active Record Inet Query Methods" do
-  before { stub_const('User', Namespaced::Record) }
+  before { stub_const("User", Namespaced::Record) }
 
   describe "#inet_contained_within" do
     let!(:local_1)     { User.create!(ip: "127.0.0.1") }